requirements.txt
streamlit
snowflake-connector-python
snowflake-snowpark-python
snowflake-ml-python